Initial commit
The project implements an additional layer store plugin for CRI-O/Podman,
so that nydus images can be mounted with lazy pulling. The plugin depends
on additional layer store plugin feature, uses go fuse implementation to
interact with podman in filesystem, and nydusd does the actual remote mount
for nydus image.

Signed-off-by: fatelei <fatelei@gmail.com>